pyproject.toml
[project]
name = "estam-clinic-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "python-decouple>=3.8",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 105
exclude = [
    "migrations",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
pythonpath = ["."]
testpaths = ["accounts/tests", "clinic/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["."]
omit = ["*/tests/*", "*/__pycache__/*", "*/htmlcov/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "model-bakery>=1.23.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
    "pre-commit>=4.3.0",
]
